+++ packages/sit-onyx/src/components/OnyxInfoTooltip/types.ts
@@ -1,4 +1,8 @@
 import type { OnyxTooltipProps } from "../OnyxTooltip/types";
 
 export type OnyxInfoTooltipProps = Required<Pick<OnyxTooltipProps, "text">> &
-  Pick<OnyxTooltipProps, "position">;
+  Pick<OnyxTooltipProps, "position"> & {
+    /**
+     * Sets the aria-label of the info icon for accessibility / screen readers
+     */ label?: string;
+  };