+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,42 @@
 # alpha-test-app
 
+## 0.1.0-alpha.125
+
+### Patch Changes
+
+- Updated dependencies [55c0b19]
+  - sit-onyx@1.0.0-alpha.128
+
+## 0.1.0-alpha.124
+
+### Patch Changes
+
+- Updated dependencies [62cb2ca]
+  - sit-onyx@1.0.0-alpha.127
+
+## 0.1.0-alpha.123
+
+### Patch Changes
+
+- Updated dependencies [85482cd]
+- Updated dependencies [edbfc22]
+- Updated dependencies [edbfc22]
+  - sit-onyx@1.0.0-alpha.126
+
+## 0.1.0-alpha.122
+
+### Patch Changes
+
+- Updated dependencies [2805e75]
+  - sit-onyx@1.0.0-alpha.125
+
+## 0.1.0-alpha.121
+
+### Patch Changes
+
+- Updated dependencies [5294c95]
+  - sit-onyx@1.0.0-alpha.124
+
 ## 0.1.0-alpha.120
 
 ### Patch Changes