+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @sit-onyx/nuxt
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- Updated dependencies [55c0b19]
+  - sit-onyx@1.0.0-alpha.128
+
+## 1.0.0-alpha.7
+
+### Patch Changes
+
+- Updated dependencies [62cb2ca]
+  - sit-onyx@1.0.0-alpha.127
+
+## 1.0.0-alpha.6
+
+### Patch Changes
+
+- Updated dependencies [85482cd]
+- Updated dependencies [edbfc22]
+- Updated dependencies [edbfc22]
+  - sit-onyx@1.0.0-alpha.126
+
+## 1.0.0-alpha.5
+
+### Patch Changes
+
+- Updated dependencies [2805e75]
+  - sit-onyx@1.0.0-alpha.125
+
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- Updated dependencies [5294c95]
+  - sit-onyx@1.0.0-alpha.124
+
 ## 1.0.0-alpha.3
 
 ### Patch Changes